fix nextjs incorrect export value error by moving fonts to a seperate file in the lib folder
