+++ src/lib/fonts.ts
@@ -1,0 +1,11 @@
+import { Inter, Bruno_Ace_SC } from "next/font/google";
+
+export const inter = Inter({ subsets: ["latin"] });
+
+export const bruno_ace_sc = Bruno_Ace_SC({
+  subsets: ["latin"],
+  weight: "400",
+  variable: "--font-bruno-ace-sc",
+  display: "swap",
+});
+
